# tests/fixtures/openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LNbits API",
    "version": "1.4.1"
  },
  "paths": {
    "/api/v1/wallet": {
      "get": {
        "tags": ["Wallet"],
        "summary": "Api Wallet",
        "operationId": "api_wallet_api_v1_wallet_get",
        "parameters": [],
        "responses": { "200": { "description": "OK" } },
        "security": [{ "APIKeyHeader": [] }]
      },
      "post": {
        "tags": ["Wallet"],
        "summary": "Create Wallet",
        "operationId": "create_wallet_api_v1_wallet_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "title": "Wallet Name" }
                },
                "required": ["name"]
              }
            }
          },
          "required": true
        },
        "responses": { "201": { "description": "Created" } },
        "security": [{ "APIKeyHeader": [] }]
      }
    },
...
    "/api/v1/payments": {
      "get": {
        "tags": ["Payments"],
        "summary": "Api Payments",
        "operationId": "api_payments_api_v1_payments_get",
        "parameters": [
          {
            "required": false,
            "schema": { "type": "integer", "title": "Limit", "default": 10 },
            "name": "limit",
            "in": "query"
          }
        ],
        "responses": { "200": { "description": "OK" } },
        "security": [{ "APIKeyHeader": [] }]
      },
      "post": {
        "tags": ["Payments"],
        "summary": "Api Payments Create",
        "operationId": "api_payments_create_api_v1_payments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvoiceData"
              }
            }
          },
          "required": true
        },
        "responses": { "201": { "description": "Created" } },
        "security": [{ "APIKeyHeader": [] }]
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateInvoiceData": {
        "type": "object",
        "properties": {
          "out": { "type": "boolean", "title": "Out", "default": false },
          "amount": { "type": "integer", "title": "Amount" },
          "memo": { "type": "string", "title": "Memo", "default": "" },
          "unit": { "type": "string", "title": "Unit", "default": "sat" }
        },
        "required": ["amount"]
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": { "type": "apiKey", "in": "header", "name": "X-API-KEY" },
      "APIKeyQuery": { "type": "apiKey", "in": "query", "name": "api_key" },
      "HTTPBearer": { "type": "http", "scheme": "bearer" },
      "OAuth2PasswordBearer": { "type": "oauth2" }
    }
  }
}